+++ Makefile
@@ -30,6 +30,9 @@ juicefs: Makefile cmd/*.go pkg/*/*.go
 juicefs.ceph: Makefile cmd/*.go pkg/*/*.go
 	go build -tags ceph -ldflags="$(LDFLAGS)"  -o juicefs.ceph ./cmd
 
+juicefs.exe: cmd/*.go pkg/*/*.go
+	GOOS=windows CGO_ENABLED=1 CC=x86_64-w64-mingw32-gcc go build -ldflags="$(LDFLAGS)" -o juicefs.exe ./cmd
+
 .PHONY: snapshot release test
 snapshot:
 	docker run --rm --privileged \
