+++ pkg/chunk/utils_windows.go
@@ -1,0 +1,34 @@
+package chunk
+
+import (
+	"os"
+	"syscall"
+	"time"
+
+	sys "golang.org/x/sys/windows"
+)
+
+func getAtime(fi os.FileInfo) time.Time {
+	stat, ok := fi.Sys().(*syscall.Win32FileAttributeData)
+	if ok {
+		return time.Unix(0, stat.LastAccessTime.Nanoseconds())
+	} else {
+		return time.Unix(0, 0)
+	}
+}
+
+func getNlink(fi os.FileInfo) int {
+	return 1
+}
+
+func getDiskUsage(path string) (uint64, uint64, uint64, uint64) {
+	var freeBytes, total, totalFree uint64
+	err := sys.GetDiskFreeSpaceEx(sys.StringToUTF16Ptr(path), &freeBytes, &total, &totalFree)
+	if err != nil {
+		logger.Errorf("GetDiskFreeSpaceEx %s: %s", path, err.Error())
+		return 1, 1, 1, 1
+	}
+	return total, freeBytes, 1, 1
+}
+
+func changeMode(dir string, st os.FileInfo, mode os.FileMode) {}
