+++ pkg/object/file_unix.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
  * JuiceFS, Copyright (C) 2020 Juicedata, Inc.
  *
