+++ pkg/object/file_windows.go
@@ -1,0 +1,15 @@
+package object
+
+import "os"
+
+func getOwnerGroup(info os.FileInfo) (string, string) {
+	return "", ""
+}
+
+func lookupUser(name string) int {
+	return 0
+}
+
+func lookupGroup(name string) int {
+	return 0
+}
