+++ pkg/utils/clock_unix.go
@@ -17,22 +17,14 @@
 
 package utils
 
-import "syscall"
+import "time"
 
-type Rusage struct {
-	syscall.Rusage
+var started = time.Now()
+
+func Now() time.Time {
+	return time.Now()
 }
 
-func (ru *Rusage) GetUtime() float64 {
-	return float64(ru.Utime.Sec) + float64(ru.Utime.Usec)/1e6
-}
-
-func (ru *Rusage) GetStime() float64 {
-	return float64(ru.Stime.Sec) + float64(ru.Stime.Usec)/1e6
-}
-
-func GetRusage() *Rusage {
-	var ru syscall.Rusage
-	_ = syscall.Getrusage(syscall.RUSAGE_SELF, &ru)
-	return &Rusage{ru}
+func Clock() time.Duration {
+	return time.Since(started)
 }
