+++ pkg/utils/logger_syslog.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
  * JuiceFS, Copyright (C) 2020 Juicedata, Inc.
  *
