+++ pkg/utils/memusage.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
  * JuiceFS, Copyright (C) 2021 Juicedata, Inc.
  *
