+++ pkg/vfs/vfs_windows.go
@@ -1,0 +1,4 @@
+package vfs
+
+const O_ACCMODE = 0xff
+const F_UNLCK = 0x01
